+++ src/app/manage/[object]/organizations-creator.tsx
@@ -63,15 +63,17 @@ export function OrganizationsCreator() {
         resetForm();
     }, [needsToRefreshForm]);
 
+    const isValidOrganizationName = name.length > 3;
+
     return (
         <form className={styles['creator-form']} onSubmit={(e) => e.preventDefault()} ref={formRef}>
             {isLoadingObject && <Loader isAbsolute={true} />}
-            <InputString label={'Name'} name={'name'} value={name} onChange={onChangeName} disabled={disabled} />
+            <InputString label={'Name'} name={'name'} value={name} onChange={onChangeName} disabled={disabled} isValid={isValidOrganizationName} isRequired={true} />
             <InputSelect label={'Type'} name={'type'} value={type} onChange={onChangeType} options={typeOptions} disabled={disabled} />
             <InputImage label={'Image'} name={'image'} image={image} onChange={onChangeImage} />
             <div className={styles['navigation-box']}>
                 {image && editParam && hasImage ? <SubmitButton label={'Delete image'} onSubmit={onDeleteImage} size="small" /> : <div></div>}
-                <SubmitButton label={editParam ? 'Edit' : 'Create new'} size="small" onSubmit={onSubmit} />
+                <SubmitButton label={editParam ? 'Edit' : 'Create new'} size="small" onSubmit={onSubmit} disabled={!isValidOrganizationName} />
                 {editParam && <SubmitButton label={'Back to creation'} size="small" onSubmit={resetForm} />}
             </div>
         </form>