+++ src/app/transactions/transaction.tsx
@@ -10,7 +10,7 @@ export function Transaction({ transaction }: TransactionProps) {
     return (
         <tr className={styles['transaction-element']}>
             <td>{new Date(transaction.date).toISOString()}</td>
-            <td>{transaction.code.expiry.toISOString()}</td>
+            <td>{new Date(transaction.code.expiry).toISOString()}</td>
             <td>{transaction.code.code}</td>
             <td>
                 <div style={{ display: 'flex', justifyContent: 'flex-end' }}>