+++ src/components/common/inputs/input-number.tsx
@@ -1,6 +1,7 @@
 'use client';
 
 import styles from '@/styles/components/common/inputs/input-string.module.scss';
+import { useState } from 'react';
 
 interface InputNumberProps {
     label: string;
@@ -9,13 +10,20 @@ interface InputNumberProps {
     value: number;
     onChange: (event: any) => void;
     min?: number;
+    isValid?: boolean;
+    isRequired?: boolean;
 }
 
-export function InputNumber({ label, name, width = '100%', value, onChange, min = 0 }: InputNumberProps) {
+export function InputNumber({ label, name, width = '100%', value, onChange, min = 0, isValid = true, isRequired = false }: InputNumberProps) {
+    const [wasEverFocused, setWasEverFocused] = useState<boolean>(false);
+    const onFocus = () => setWasEverFocused(true);
+
+    const isDataValid = !wasEverFocused || isValid;
+
     return (
         <div className={styles['input-string']}>
-            <label htmlFor={name}>{label}:</label>
-            <input type="number" name={name} style={{ width }} value={value} onChange={onChange} min={min} />
+            <label htmlFor={name}>{label}: {isRequired ? '(*)' : ''}</label>
+            <input type="number" name={name} style={{ width }} value={value} onChange={onChange} min={min} onFocus={onFocus} data-is-valid={isDataValid} />
         </div>
     );
 }