+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.36.0] - 2024-06-01
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): Checking validity of inputs' form.
+- [Marek Kurańda](https://github.com/mjkuranda): Links to `Categories` and `Shops` at `TopBar`.
+
+### Changed
+- [Marek Kurańda](https://github.com/mjkuranda): Visibility of transaction table when no transactions.
+
 ## [0.35.0] - 2024-05-30
 ### Added
 - [Kamil Wojciechowski](https://github.com/kmlwoj): Dockerfile definition.